LayoutTests/fast/events/message-port-inactive-document-expected.txt

CONSOLE MESSAGE: line 31: ReferenceError: Can't find variable: MessageChannel
FAIL: Timed out waiting for notifyDone to be called
Test that messaging an inactive frameless document works as expected.
